functions/data_calls/requirements.txt
attrs==19.3.0
boto3==1.9.228
botocore==1.12.228
certifi==2019.6.16
chardet==3.0.4
coverage==5.2
docutils==0.15.2
Faker==2.0.1
idna==2.8
importlib-metadata==1.7.0
ipaddress==1.0.22
jmespath==0.9.4
kafka-python==1.4.7
MonthDelta==0.9.1
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==0.25.3
pluggy==0.13.1
psycopg2==2.8.3
py==1.9.0
pycrypto==2.6.1
pydash==4.7.5
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.0
python-dotenv==0.14.0
python-snappy==0.5.4
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.6
text-unidecode==1.2
urllib3==1.25.3
pyyaml==5.2
wcwidth==0.2.5
xmltodict==0.12.0
zipp==3.1.0
